+++ gaseio/__init__.py
@@ -8,7 +8,7 @@ from .main import read_preview, write_preview, preview, get_write_content
 from .main import list_supported_write_formats
 
 
-__version__ = '2.8.0'
+__version__ = '3.0.0'
 
 
 def version():